Updated classInfo function
Implemented prior functionality from within a while loop.

In this implementation, the desired fields are written as strings to a
list and printed.
NOTE TO SELF: Depending on the shape the rest of the program takes,
this list will either have to be returned from this function to be
formatted and printed elsewhere -or- printed from within the function
in the required output format.