Updated studentScores function
Began work on recognizing where to find the relevant data in the file.
The casefold() method forces all chars in a string to lowercase.
Since both "Name" and "SSN" should always appear in the column headers
of the input file but because one of them may appear in another line,
the check for the correct line includes both of them.