Fixed studentScores // Opt & Post sections
***I don't know why the original conditions did not work. Bring this up
to Prof.***
Workaround successful:
Changed conditions of first if to look for the string 'F' in the aopt
string. 'F' since of the three possibilities for this field, it is
contained in both "A-F" and "P/F", but not "AUD" which had tested false
positive previously. To be sure aopt contains the string "AUD", the
second if checks for the string "UD".