Added nested function gradingInfo
With the grades pulled from the file and added to the agradeList list,
this function first matches each item (grade) to its appropriate type
by comparing its position with that of its equivalent header from the
headerList. Then using the parameters for grading set out in the
gradingInfo function, it totals and processes the grades.